+++ packages/synapse-interface/utils/route/bridge/getToChainIds.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash'
 
 import { EXISTING_BRIDGE_ROUTES } from '@/constants/existingBridgeRoutes'
-import { RouteQueryFields } from './generateRoutePossibilities'
+import type { RouteQueryFields } from '@/utils/route/types'
 import { getTokenAndChainId } from '@utils/route/getTokenAndChainId'
 
 export const getAllToChainIds = () => {
