--- packages/synapse-interface/utils/route/bridge/getTokenAndChainId.ts
@@ -1,5 +1,0 @@
-export const getTokenAndChainId = (tokenAndChainId: string) => {
-  const [symbol, chainId] = tokenAndChainId.split('-')
-
-  return { symbol, chainId: Number(chainId) }
-}
