+++ packages/synapse-interface/utils/route/types.ts
@@ -1,0 +1,6 @@
+export interface RouteQueryFields {
+  fromChainId?: number
+  fromTokenRouteSymbol?: string
+  toChainId?: number
+  toTokenRouteSymbol?: string
+}
